refactor(story-close): collapse the three format-autofix modules into one (refs #4017)
